Restructure app for maintainability
Fixes #14

Restructure the app to improve maintainability by moving each component or class to its own file.

* **MainScreen**: Create `lib/screens/main_screen.dart` and move the `MainScreen` class from `lib/main.dart` to this file. Import necessary packages.
* **ConfigurationScreen**: Create `lib/screens/configuration_screen.dart` and move the `ConfigurationScreen` and `_ConfigurationScreenState` classes from `lib/main.dart` to this file. Import necessary packages.
* **Main.dart**: Remove the `MainScreen`, `ConfigurationScreen`, and `_ConfigurationScreenState` classes. Import `lib/screens/main_screen.dart`.
* **BackgroundService**: Move the `_showSystemAlert` method from `MainActivity` to `lib/services/background_service.dart`. Implement the `_showSystemAlert` method to show a system alert. Add methods `_checkOverlayPermission` and `_showLockDialog`.
* **MainActivity**: Remove the `_showSystemAlert` method and call the `_showSystemAlert` method from `BackgroundService` instead.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/bertusviljoen/flutter_time_lock/issues/14?shareId=XXXX-XXXX-XXXX-XXXX).
