+++ lib/main.dart
@@ -3,6 +3,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_background/flutter_background.dart';
 import 'configuration.dart';
 import 'services/background_service.dart';
+import 'screens/main_screen.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -22,192 +23,6 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       home: MainScreen(),
-    );
-  }
-}
-
-class MainScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Main Screen'),
-      ),
-      body: Center(
-        child: ElevatedButton(
-          onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => ConfigurationScreen()),
-            );
-          },
-          child: Text('Go to Configuration'),
-        ),
-      ),
-    );
-  }
-}
-
-class ConfigurationScreen extends StatefulWidget {
-  @override
-  _ConfigurationScreenState createState() => _ConfigurationScreenState();
-}
-
-class _ConfigurationScreenState extends State<ConfigurationScreen> {
-  static const platform = MethodChannel('com.example.flutter_time_lock/system');
-  Map<String, dynamic> config = Configuration.config;
-
-  final _formKey = GlobalKey<FormState>();
-
-  @override
-  void initState() {
-    super.initState();
-    _checkOverlayPermission();
-    _startBackgroundService();
-  }
-
-  Future<void> _checkOverlayPermission() async {
-    try {
-      bool hasPermission =
-          await platform.invokeMethod('checkOverlayPermission');
-      if (!hasPermission) {
-        await platform.invokeMethod('requestOverlayPermission');
-      }
-    } catch (e) {
-      print('Error checking overlay permission: $e');
-    }
-  }
-
-  Future<void> _startBackgroundService() async {
-    await BackgroundService.startService(config);
-  }
-
-  Future<void> _showSystemAlert() async {
-    try {
-      await platform.invokeMethod('showSystemAlert',
-          {'title': 'Lock Alert', 'message': 'Time to lock the device!'});
-    } on PlatformException catch (e) {
-      print("Failed to show system alert: ${e.message}");
-      // If permission is denied, request it again
-      if (e.code == 'PERMISSION_DENIED') {
-        await _checkOverlayPermission();
-      }
-    }
-  }
-
-  void _showLockDialog() {
-    _showSystemAlert(); // Always use system alert instead of Flutter dialog
-  }
-
-  Future<void> _saveConfig() async {
-    if (_formKey.currentState!.validate()) {
-      _formKey.currentState!.save();
-      await Configuration.saveConfig(config);
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Configuration saved')),
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Flutter Time Lock'),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Form(
-          key: _formKey,
-          child: Column(
-            children: <Widget>[
-              TextFormField(
-                initialValue: config['childPin'],
-                decoration: InputDecoration(labelText: 'Child Pin'),
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return 'Please enter a child pin';
-                  }
-                  return null;
-                },
-                onSaved: (value) {
-                  config['childPin'] = value!;
-                },
-              ),
-              TextFormField(
-                initialValue: config['adultPin'],
-                decoration: InputDecoration(labelText: 'Adult Pin'),
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return 'Please enter an adult pin';
-                  }
-                  return null;
-                },
-                onSaved: (value) {
-                  config['adultPin'] = value!;
-                },
-              ),
-              TextFormField(
-                initialValue: config['lockTime'],
-                decoration: InputDecoration(labelText: 'Lock Time'),
-                keyboardType: TextInputType.number,
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return 'Please enter a lock time';
-                  }
-                  if (int.tryParse(value) == null) {
-                    return 'Please enter a valid number';
-                  }
-                  return null;
-                },
-                onSaved: (value) {
-                  config['lockTime'] = value!;
-                },
-              ),
-              TextFormField(
-                initialValue: config['lockInterval'],
-                decoration: InputDecoration(labelText: 'Lock Interval'),
-                keyboardType: TextInputType.number,
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return 'Please enter a lock interval';
-                  }
-                  if (int.tryParse(value) == null) {
-                    return 'Please enter a valid number';
-                  }
-                  return null;
-                },
-                onSaved: (value) {
-                  config['lockInterval'] = value!;
-                },
-              ),
-              DropdownButtonFormField<String>(
-                value: config['timerUnit'],
-                decoration: InputDecoration(labelText: 'Timer Unit'),
-                items: ['seconds', 'minutes']
-                    .map((unit) => DropdownMenuItem(
-                          value: unit,
-                          child: Text(unit),
-                        ))
-                    .toList(),
-                onChanged: (value) {
-                  setState(() {
-                    config['timerUnit'] = value!;
-                  });
-                },
-                onSaved: (value) {
-                  config['timerUnit'] = value!;
-                },
-              ),
-              SizedBox(height: 20),
-              ElevatedButton(
-                onPressed: _saveConfig,
-                child: Text('Save'),
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
